Decoupling Authentication and Loading in various controller hierarchy
